Add custom monetary classes
The Bitcoinj monetary classes do not support base and quote currency but are centered around Coin and Fiat making it less flexible for fiat-fiat or crypto-crypto where base and quote currency can be defined by the user. Also the api is heavily overloaded and the exchangerate class is not intuitive to use. It further misses a currency code for the coin class (which is designed to be BTC).

I added a Quote class which represents the price. Also added formatting so no need for an external formatter.

I leave the bitcoinj package still there until the new classed are more tested and applied.