Add lombok dependencies
I am aware that buildSrc is more for common tasks instead of sharing dependencies, but as we use it for logging as well I think it is convenient to repeat that pattern for application wide used dependencies. At least I don't see any downside...

After some doubts I tend to be in favor of adding lombok as it improves productivity and reduces boilerplate code and thus increase readability.
Java records are not flexible enough to replace all use cases for value objects, as well as other annotations are convenient as well.