Add common module
This is a collection of used utils and common classes from the misq prototype.
It has only few tests so far. Many classes are copied or derived from Bisq. There might be some tests for those to take over.

The monetary package was created in favor over the BitcoinJ monetary classes as those have heavily overloaded APIs and do not cover the use cases we need.
They are centered around BTC but we need to support fiat-fiat as well as altcoin-altcoin pairs.
Beside that the exchange rate API was very un-intuitive to use and did not cover all use cases.
The Quote class should handle all the use cases we need and was used already in the prototype project for the offer book domain.
Handling of BigInteger and BigDecimal might need more tests and checks if it handles overflow scenarios correctly.