Add i2p module
Working code using SAM protocol. We might consider to use I2P java library classes instead but that way we do not inherit lot of dependencies so for the moment it is the preferred way. Not intensively tested yet but in for development testing. I2P need to be running and it takes about 1- 2 minutes before its ready to accept SAM connections.